SesizariController.php Restricted only for admin

+++ frontend/controllers/SesizareController.php
@@ -37,13 +37,16 @@ class SesizareController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new SesizareSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin')){
+            $searchModel = new SesizareSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+        return  $this->redirect(['site/index']);
     }
 
     /**
@@ -54,9 +57,13 @@ class SesizareController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin')){
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
+        return  $this->redirect(['site/index']);
+
     }
 
     /**
@@ -90,15 +97,19 @@ class SesizareController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin')){
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
+        return  $this->redirect(['site/index']);
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -110,9 +121,12 @@ class SesizareController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin')){
+            $this->findModel($id)->delete();
+            return $this->redirect(['index']);
+        }
+        return  $this->redirect(['site/index']);
 
-        return $this->redirect(['index']);
     }
 
     /**
